CODETOOLS-7903320: Move InvokeMethodAdapter.getInvokeID(String, String, String) up
